Remove an unnecessary helper method to avoid confusion.
Internally, the word "encode" is a bit overloaded: we use it to mean
plaintext data that has been compressed and then encrypted against the
pad, but we're also doing Base64 encoding.  Having the Base64 encoding
happen in a separate function just increased the chance of confusion.
After this change, "encode" really means all the steps, including the
Base64 encoding, and is once again symmetrical with the decoder.

* onetime
  (PadEncoder._output): Remove this method.
  (PadEncoder.encode, PadEncoder.finish): Adjust accordingly.