Implement tail fuzz message authentication (#2)
We already authenticated the message using head fuzz.  Now check the
tail fuzz too, which means that even if the message decrypted
successfully, there will still be an authentication error if the tail
fuzz was tampered with.  This commit also fixes the decryption-case
pad-record usage bug for which a test was added in commit 9e8422f07.

* onetime: Bump version to 2.0-beta6.
  (PadSession.convert): Add comment explaining tail fuzz reserve.
  (PadSession.finish): Verify tail fuzz.
  (PadSessionDecoder.finalize): Remove now-obsolete comment.

* check.sh
  ("decode v2 msg, where v1 entry has range already used",
   "decode v2 msg, where v1 entry range needs stretching",
   "decode v2 msg, where v1 entry needs new range"): Adjust numbers.
  ("decryption should record same pad usage as encryption"): Expect pass.
  ("tamper with tail fuzz to cause authentication error"): New test.