Start on combined digest.
* onetime
  (PadSession.__init__): Rename _head_fuzz_digest to _msg_hash.
  (PadSession.digest_gulp): New method.
  (PadSession._verify_digests): Rename to...
  (PadSession._verify_digest): ...here, and adjust.
  (PadSession._consume_fuzz_bytes, PadSession._make_fuzz): Just take
    a head fuzz flag instead of a hash object, and conditionally
    update the internal session hash based on the flag.
  (PadSession.finish): Adjust.

Add a lot of debugging prints to discover that we can't count on
convert() getting called (and thus the session has getting initialized
with the head fuzz) on the first call to encode(), because the
compressor might not return anything on a given call and in particular
usually doesn't on the first call.  IOW, this doesn't work yet.