Start debugging an interesting behavior.
While fixing the tests, I found an unexpected newline in the middle of
the base64-encoded region of an encrypted output file.  That led to
the realization that the deconditionalization of the convert call(),
in commit d7b8d9eb, had changed the boundaries of the base64 output.
This doesn't necessarily affect the correctness of the output, but it
does affect the aesthetics, and it produces an artifact that will
interfere with future debugging.  However, getting rid of that newline
is turning out to be surprisingly complicated; this commit is a start.

* onetime
  (SessionEncoder.__init__): New _started_encoding flag.
  (SessionEncoder.encode): Use the flag.  Emit debugging information.

* check.sh: Adjust a bunch of tests.
  ("tampering with ciphertext causes bzip decoder error"):
  Add exits around current debugging.